Continuing development
Changed Verbosity Level to a bitmask to allow more control over output
messages.

Added the concept of a command modifier.   Commands generally take the
form of <noun><verb><modifier> <argument>.  For example, NV 1 sets the
Navigator's message mask to 1; NV+ 1 adds the 1 bit to the existing
message mask; NV- 1 removes the 1 bit.

Added the 'Q' command to the common commands handled by the Actor base
class.  This is to be used to Query current parameters of each Actor.
Common values, such as enablement and verbosity, are handled by the
Actor base class.  Specific values are handled by the derived class.