Menu mode changes
Changed CommandDispatcher to implement a simple "menu mode".  To enter
this mode, just type the initial letter (the Command letter) of the
CommandSubscriber, with no subcommands or arguments.  This will cause
that object's menu to be displayed, and subcommands from that menu can
be chosen.  Entering a blank commandline will return to the top-level
menu.

This does not change the previous commandline options, it works on top
of them.