LEDDriver Changes
Changed LED Driver to work with bi-polar two-pin LED's.  This makes the
operation closer to how real motor drivers would work.  That is, each
motor would be controlled by one PWM output and one direction output.

Added a helper function to set the led outputs.  This removes some
redundancy.

Changed the sketch to use the new pins.  The actual pins were chosen for
convenient connection using an Arduino Pro Mini on a breadboard.  The
two end pins (9 & 10) are PWM pins, connected directly to the left and
right LED's.  The other sides of the LED's are connected through
resistors to pins 5 & 14, chosen because that's where the resistor
reached easily.